move to konnektor.scoring
