+++ src/konnektor/network_planners/concatenators/cyclic_concatenator.py
@@ -7,7 +7,7 @@ from collections.abc import Iterable
 from gufe import AtomMapper, LigandNetwork
 
 from .._networkx_implementations import MstNetworkAlgorithm
-from ..scorer import AtomMappingScorer
+from ..scoring import AtomMappingScorer
 from ._abstract_network_concatenator import NetworkConcatenator
 from .max_concatenator import MaxConcatenator
