+++ src/konnektor/network_planners/generators/explicit_network_generator.py
@@ -8,7 +8,7 @@ from collections.abc import Iterable
 
 from gufe import AtomMapper, Component, LigandNetwork
 
-from konnektor.network_planners.scorer import AtomMappingScorer
+from konnektor.scoring import AtomMappingScorer
 
 from .._map_scoring import _parallel_map_scoring
 from ._abstract_network_generator import NetworkGenerator
