+++ src/konnektor/network_planners/generators/n_node_edges_network_generator.py
@@ -8,7 +8,7 @@ from gufe import AtomMapper, Component, LigandNetwork
 from konnektor.network_planners._networkx_implementations import (
     NNodeEdgesNetworkAlgorithm,
 )
-from konnektor.network_planners.scorer import AtomMappingScorer
+from konnektor.scoring import AtomMappingScorer
 
 from ._abstract_network_generator import NetworkGenerator
 from .maximal_network_generator import MaximalNetworkGenerator
