+++ src/konnektor/network_planners/generators/twin_star_network_generator.py
@@ -6,7 +6,7 @@ from collections.abc import Iterable
 from gufe import AtomMapper, Component, LigandNetwork
 
 from konnektor.network_planners._networkx_implementations import RadialNetworkAlgorithm
-from konnektor.network_planners.scorer import AtomMappingScorer
+from konnektor.scoring import AtomMappingScorer
 
 from ._abstract_network_generator import NetworkGenerator
 from .maximal_network_generator import MaximalNetworkGenerator
